Move API to a separate Django app
