+++ api/tests/test_api.py
@@ -26,12 +26,12 @@ import os
 
 from django.test import TestCase
 
+from api.serializers import PackageSerializer
 from core.models import Package
 from importers.data_dump import debian_dump
 from importers.data_dump import ubuntu_dump
 from importers import debian
 from importers import ubuntu
-from vulnerabilities.api import PackageSerializer
 
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
