+++ api/urls.py
@@ -25,7 +25,7 @@ from django.conf.urls import url
 
 from rest_framework.urlpatterns import format_suffix_patterns
 
-from vulnerabilities import views
+from api import views
 
 
 urlpatterns = [
