+++ vulnerablecode/settings.py
@@ -46,9 +46,9 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'django_filters',
+    'api',
     'core',
     'importers',
-    'vulnerabilities',
 ]
 
 MIDDLEWARE = [
